Added car-script
Still looking for a way to slow down the speed when the throttle (or brake) isn't pressed. Currently it retains it's last speed when driving straight, even if the accelerate button isn't pressed. Already tried by setting brake value to 1, but that interferes with the rest of the setup with breaking and reversing.
Want to find a solution that brings down the velocity to 0 at a certain pace.

On the other end, car looses too much speed when turning the vehicle.